fix: repair broken autofix decompositions — imports, module names, doc comment placement
Fix all issues introduced by CI autofix decomposition commits:

move_items decomposition:
- Rename whole-file_move.rs → whole_file_move.rs (hyphens invalid in Rust modules)
- Move doc comments above mod declarations (inner doc comments can't follow items)
- Fix decompose.rs: use super::move_items::{MoveOptions, MoveResult} instead of
  super::move_options (module was reorganized)
- Fix extension_integration.rs: replace broken glob with specific imports
- Fix item_kind.rs, move_options.rs: remove unnecessary broken glob imports
- Fix types.rs: use super::ItemKind instead of broken glob
- Fix whole_file_move.rs: import core_parse_items, ImportRewrite from super

version decomposition:
- Fix default_pattern_for_file.rs: add local_files + FileSystem trait import,
  use super:: paths instead of absolute crate::core::release::version:: paths
- Fix version.rs: use super::read_local_version, remove broken glob
