+++ src/core/refactor/move_items/item_kind.rs
@@ -1,7 +1,5 @@
 //! item_kind — extracted from move_items.rs.
 
-use crate::core::refactor::*;
-
 
 #[derive(Debug, Clone, serde::Serialize)]
 #[serde(rename_all = "snake_case")]
